Update file naming throughout book
Since the order of the chapters isn't definite yet, I think the renaming
of the chapter files was going to get really confusing in the version
control. Naming the files by their title still isn't a perfect solution
because it will require a handful of changes when renaming, but this
will prevent confusion where GitHub would essentially show a diff of a
chapter being changed to completely different content.