Initial rewriting for introduction.
Finally getting a chance to start rewriting content from the ground up.
Trying to remove some of the extraneous formatting that was meant for
the web marketing copy, but probably not very necessary in a book
format. Hoping to continually rewrite and tinker with the content until
arriving at something markedly better.