Resolve proselint suggestions
Ended up having some trouble getting proselint to work the way I was
hoping for. I had to symlink the .proselintrc file to my home directory
for the command-line utility to work. Ideally I'd like to keep the
symbol checks in and just ignore things like ellipses and curly quotes
that I think are bad ideas.

For now, I'll keep running proselint manually to get suggestions here
and there, but going to hold off on automating it since I intend to
ignore some of the checks.