Configuration for Game.elm file
Seemed to take me forever to get this configured properly. This was
supposed to be the easy part in regards to having a dedicated game file
to work from. Still not sure how to make this flexible yet and adapt to
different game titles with different Elm files. My hope was always that
each game would be a separate Elm source code file, and that would make
each game completely self-contained. But I'll have to look into the best
way to handle this. For now this is good enough to start adding a game
here and working towards tracking a score.