fix Gameplay relationships
Prior to this commit Gameplay specified that is `has_one` Player and
`has_one` Game, which imply that both Player and Game have a
`:gameplay_id` field representing the foreign key. This indicates the
relationship is backward and Gameplay `belongs_to` Game and `belongs_to`
Player, which uses the underlying `:game_id` and `:player_id` we've
already created. This commit corrects the directionality of the
relationship and updates the link to the documentation for further
reading.