Update initial installation instructions
I still think it's a good idea to keep the installation information
consolidated to a page in the back of the book. Hopefully this will make
it relatively easy to keep updating the book to the latest versions
without having to update the intro chapters individually.

I also pared down the callout for the Phoenix 1.3 info. It may not be
necessary anymore now that Phoenix 1.3 is standard, but that was a big
point of confusion for many who had previously seen intro material based
on 1.2.