Add container div to reduce minor confusion
Without the container div the buttons end up collapsing after converting
from HTML to Elixir code. It's not a big deal, but it was slightly odd
to see a change when we were essentially accomplishing the same thing.
The container div should hopefully reduce some mental strain or going
off on a tangent to try and figure out what's happening like I did the
first time I tried this.

I've also been debating between using `link` and `link()`, but I tested
the new Elixir 1.6 formatter and I think the parens are encouraged. So
I'll try to be consistent about using them where possible.