Remove section on database seeds
I'd really love to keep this content, but it does create quite a bit of
confusion, and this chapter is already really long to work through.

The issue with the database seeds in the existing format is that using
Repo.insert! will work, but any accounts created in this manner will not
be usable do to the missing password hashing. And while the
`create_player` function works, I don't think this is a common practice
for creating demo seeds. And it doesn't feel right seeding data with
password fields in the first place.