Rewording in attempt to fix PDF issue
It looks like the URL near the beginning of this paragraph was causing
some problems in the PDF. It was extending all the way to the right of
the page outside the readable bounds. Hopefully the slight rewording
will adjust it enough to have all the content fit properly in the PDF
output.