Temporary removal for gameplay chapters
I'm working towards a more stable version of the book that doesn't have
any half-written chapters. The idea here is that the minigame has just
enough features to switch gears and work on syncing scores. I consider
that a must for the book to be released, whereas additional game
features would be nice but not an utter necessity since the game has a
working scoring mechanism at this point in the book.

The game states chapter will now transition into the syncing score data
chapter, and then I can add in additional game chapters as the content
is stable.