Trim initial output for Phoenix project creation
Seemed a bit odd to have such a huge amount of output that took up a
full page of book content. It's still a lot even trimmed, but I think
this might be a better idea in terms of getting a good idea of the
folder structure without overwhelming with specific filenames.