Rewrite Gameplays section of API chapter
I actually preferred the manual file creation approach for these
features, but later in the book it would require a _ton_ of extra
content to add the controller. The original way worked well because we
manually created the schema to see the relationships. But this way
allows us to get all the benefits that the generators provide, and make
it way easier to query for gameplays later in the book.