Game setup WIP
Struggling to figure out why the Platformer.elm file won't load when the
Main.elm file seems to work. I think I've narrowed it down since the
controller, route, and template all seem to work okay. It has to be
something with app.js, webpack.config.js, or Platformer.elm (or some
combination of them). Committing this content for now so I can keep
moving and circle back to the configuration.