Update progress, finishing with Channel chapter
There was enough thrashing in this chapter that I'll have to run through
it again from scratch to make sure everything works as advertised.
Thankfully I'm nearing the end with all the rewrites, and I'll replace
the original sample app with a new version as I verify all the updated
content works well.